pyproject.toml
[project]
name = "scholarps-exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "prometheus-client>=0.23.1",
    "pyvirtualdisplay>=3.0",
    "pyyaml>=6.0.3",
    "selenium>=4.34.2",
]
